Updates and fixes alchemy workshop/clinic garden (#3854)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<img width="506" height="275" alt="image"
src="https://github.com/user-attachments/assets/1e9e99e7-fd8d-4c9f-9b56-2ed67f028545"
/>
Replaces the brazier with a sconce and removes the weird stone wall that
was above one of the plots covering it from sunlight

## Why It's Good For The Game

Gnomes were reported to get stuck on the brazier. And the floating wall
caused plots to be unable to grow.
view from the top Z level
<img width="551" height="579" alt="image"
src="https://github.com/user-attachments/assets/325c7be0-255f-4c21-890e-f48ec69736cb"
/>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
map: removed a brazier in the apothecary garden that caused gnomes to
get stuck on
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->