Storyteller Maintenance (#3879)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Maintenance PR.

Changes:

- Tweaks default point threshold values to match live server
- Slightly higher storyteller point gain in general
- God intervention events now use more law list of allowed storytellers,
usually based on divine and inhumen pantheon
- Ascendant gods will only use general and events dedicated to them
- Lesser point penalty for failing events
- Longer Week communal cost 80 -> 70
- Some personal events are easier
- New storyteller tags
- Influences and storyteller tags balancing
- New Graggar influence - number of hunted people
- Intervention personal objectives escalate only intervention track
- No triumph bonus modifer when admins give triumphs or for psydon's
fund
- Hide admin rankers from triumph leaderboards
- Small tweaks and cleanup to intervention and personal
events/objectives

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->