Trims down the dwarf accent (#3831)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes all of the replacements that are just flat out a different
language/sound nothing like what you're trying to say.
**This is NOT removing most of the accent** this is just taking away the
conversions that someone looking at wouldn't be able to figure out even
with a few seconds of looking at a dwarf's speech
like 
stunned : donnert
or 
frown : glunch

### Note: If there's any specific things that should stay, let me know.

## Why It's Good For The Game
While the dwarf accent is funny, it's also an enourmous pain in the ass
to try and understand
for example, guess what the word for shy was, 
just guess.
the correct answer was **blate**
I have seen so, so many dwarves have to say in Looc "this is what I
meant:"
The accent is funny, but god it is a pain in the ass to play and try to
speak to people with.
I really don't think having to ask in Looc "what did you say" or look up
whatever random word like donnert online just to understand what john
dwarf says is particularily good for the game.
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
del: removes a lot of the godawful replacements in the dwarf accent
qol: makes dwarves actually a bit comprehendable
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->