Storyteller Maintenance III (#3931)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

With some extra misc small fixes.

Changes:

- Notable people now receive 1 triumph at round end
- Fixes Abyssor's new release fish objective
- Antag events tweaks
- Fix for werevolf spread objective
- Leftover tweaks and fixes for inquisitor roles seperation from the
church
- Cloth pouch now salvages into cloth, not leather
- Changed a few rot timer to use defines, and increased it for eggs
- Influences balancing

Fixes #3922

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->