(Bounty) Tomb of Matthios Pit Tiles (#3912)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This adds a new turf - `/turf/open/dungeon_trap` When stepped onto, you
will fall into a random area of the Matthios Dungeon if it is loaded.
This is a one way trip, you can't go back the way you came, and you
almost certainly won't be dropping together with someone else.

The algorithm to find where to drop you goes vaguely:
- Gather all top level tiles from the dungeon and pick one
- If it is an open space, drop there. If it is a closed space, check the
space below it to see if you could drop there.
- If it didn't find one or if there is a dense, non-climbable structure
in the way, try again.

This also adds it as implementation into the Vanderlin keep.

This will likely softlock at least a few people in the event you get
dropped into a closed off room. They aren't too common but it is
possible.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Ook bounty for pit tile

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: New Chasm tiles that drop you into a random part of the Tomb of
Matthios with more than likely a few broken limbs.
map: New area to Vanderlin Keep next to Treasury.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->