Fix and Changes Class Select Handler (#3873)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Fix Assassin not always appearring

- Rare Classes are outside the normal class limit, meaning if you
sucessfully roll them it will appears even if the amount of classes on
the menu are maxed (for example in the case of 15 non rare classes for
the pilgrim, rare classes will still appear)

- Shuffles the normal list of classes, this means if your job got a
limit of 5 normal classes it won't always be the same 5 classes.

Closes #3847 

## Why It's Good For The Game

Fix.

Rare Classes were actually pushing normal classes, this means you could
roll something rare and it would push some other role that you could
have wanted to play.

This affects specifically more like migrant waves that got a lower limit
for the same jobs (pilgrim and adventurer) and it would just appear the
same 5 or 10 classes instead of randomizing between the non rare.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Assassin not appearing as pilgrim
code: Rare Classes does not count toward the limit of classes in the
class menu.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->