Flammability, Soaking, and Soap (#3884)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- This adds the ability to wash off fire stacks by several means. These
would include washing by right click, walking into water, or using soap
in water (though you'll probably be cleaned already). This operates
through the `SoakMob()` proc in living.dm
- When someone is on fire, inspecting them will no longer say they are
covered in something flammable if they are on fire. It will say they are
on fire.
- Pyromaniacs can sate their addiction from examining someone on fire
rather than being close to them.
- Sitting in water will slowly make you wet. I haven't set this up for
getting splashed with water since that would need to make it a touch
effect for a lot of the non-alcoholic reagents.
- There's a tiny little refactor of a redundant hygiene code check.
- Since it wasn't being used, the herbal soap buff no longer needs to be
applied by another person, but has had its duration lowered as a result.
- Negative fire stacks (being soaked) goes away at 40% it's previous
rate.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

- There was no way to get rid of flammability unless you set yourself on
fire and extinguished yourself, as flammability only went away when the
Extinguish proc successfully called.
- Being soaked went away fast enough to be inconsequential, this'll have
it last 2 or so minutes before you fully dry from fully wet.
- The initial philosophy behind having someone else apply the soap buff
was somewhat of a layover from the bathmaid code. I assumed this would
be a fun royalty roleplay, being bathed by their servants.
Unfortunately, no one is doing this. People are either unaware of its
existence or assumed it had a different intent. At this point I think
it's best if you can apply it yourself, albeit for only 10 minutes
instead of 15.
 
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: You can now wash off flammables in water by washing yourself or
entering it for a period of time.
add: Water will now soak you as you remain in it. Soaking prevents fire
but will dry off after a while.
add: Pyromaniacs can now sate their addiction by examining someone who
is on fire rather than needing to be close to them. This doesn't apply
to non-mobs.
balance: Being soaked dissipates at 40% the rate it previously did.
balance: Removed the need to be scrubbed by someone else when using
herbal soap to acquire the buff, but shortened it's duration to 10
minutes instead of 15.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->