Hidden Door Tweaks and Rebalancing (#3878)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This is a variety of fixes/changes to how hidden doors work. This mostly
pertains to how to detect them and using perception to allow hidden
doors to be seen/used. The most important takeaway is that it isn't a
hard restriction to know the plans or the password to use the doors.

## Why It's Good For The Game

I've noticed they've become heavily underutilized since #1012. This aims
to help steer them more towards a usable feature while keeping some
level of meta protection in place.


## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Hidden levers can be detected by looking around, similar to a trap.
Only visible to those who meet the perception DC of the lever.
qol: Roles that know the keep's hidden door passwords now have it added
to their memories.
balance: Hidden doors and levers now have a perception DC to be used. A
standard hidden door/lever has a DC of 10. A keep door/lever is 14.
Thieves Guild members and Assassins get a +2 bonus to this check.
map: Re-added Keep architectural plans to Vanderlin's archive room.
These are used to reveal the locations of the hidden doors.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>