(WRETCH) Bloodsucker/Renegade Changes/Additions/Adjustments (#4060)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Adds carpentry/crafting to the Bloodsucker subclasses. They can't
craft a coffin to rest in without it.

- New Bloodsucker Subclass (Wretch) - The Vagrant. Drifter reference?!?!
Unarmed/Dodge. This might be busted, but fuck it for now if it becomes a
problem we can punch it down later.

- Adds an optional list of titles for the Bloodsucker.

- Makes the Bum subclass for Bloodsucker less miserable. Minimum of 7
luck, minimum of skills, minimum of stats.

- 1-point combatskill buffs for Bloodsucker subclasses.

- Adds more titles to the Renegade.

- Adds dodge to the Renegade, skill adjustment too. He is MISERABLE
without any defense skills. I almost died to an NPC bum even after
landing a puffer shot, he sucks that bad.

- Adds knife-skill to the Renegade. 2 minute cooldown for a weapon that
can easily miss, cannot Point-Blank anymore, and doesn't guarantee a
pain-crit. If you can't remove someone with one shot you need a
fallback. He's getting a weaponskill. Does not help that there's
currently a bug that kills the gun if you don't pop a shot off with
left-click.

- Adds the song Last Day As A Man (First Night As A Beast) by
Burialgoods - https://youtu.be/zznEa8EhpuQ - Gives it to Renegade Wretch
for now. I'd give it to Bloodsucker but there's an antag vampire
overwrite I don't know how to work around, I just want it in the code
right now, yes the song was requested by Ook awhile ago.

- Increases rollchance to 100% for Vigilante, adds total positions
(**NOT** JOB, SUBCLASS) to 10 for Bloodsucker and Vigilante - Wretch is
already a rare chance to get since a fair number of people que for it,
we shouldn't have there be a chance to only get one instance of these
classes. Also, subclass shit is wonky and granting people a specific
subclass is (apparently) overly complicated. I don't like "uhmm in my
opinion" changes but we probably shouldn't have a maximum for these
subclasses, better to lessen headaches and let people have fun. Players
should come first here.



## Why It's Good For The Game

These subclasses are miserable to play and needed touched-up. Also, new
shit, new gimmicks, new fun. Behold fun.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->