(SOUND) Psydonite Musicbox gets a new sound (#4076)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

It gets a new sound. This time it's not lifted from fucking
Phasmophobia.

## Why It's Good For The Game

Revelation 21:8
"But the cowardly, the unbelieving, the vile, the murderers, the
immoral, those who practice magic arts, the idolaters and all liars—they
will be consigned to the fiery lake of burning sulfur. This is the
second death.”


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->