(Bounty) The Glutton (#3894)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- The Glutton is a new special trait for Monarch with a weight of 20.
- Their stats are adjusted to gain 5 STR, 7 CON, and lose 6 END and 10
SPD.
- They gain 3 in wrestling (legendary) but lose 3 in athletics (weak).
- They gain crit resistance because they're big boned.
- Added usage to the fat trait. This is now visible on a player similar
to being beautiful, but has no other effects as of now. Could see use
for object interactions?
- The Glutton gets scaled to be a little wider.
- The Glutton has certain noises replaced with other ones for comedic
effect. They also gain a unique noise called the Gluttonous Burp. It is
louder and can be heard from further away.
- **Most importantly**, the Glutton and their entire family have the
curse of Insatiable Hunger upon their family tree. Insatiable Hunger
increases your nutrition mod to by 1.1, to a total of 2.1, which is 0.1
more than a Half-Orc. If the Baothan curse is lifted, this will return
to its previous value. The Royal Family will not receive any of the
benefits of the Monarch.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

It's funny, there's tons of fat kings who play so this would be a fun
one for them.
Ook also turned this into a bounty

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: New Special roll: The Glutton. Monarch only... and their family.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->